requirements.txt
fastapi~=0.115.6
fast_auth_tools==0.2.2
uvicorn~=0.34.0
starlette~=0.41.3
aiosqlite~=0.20.0